requirements.txt
python-dotenv
psycopg2
apache-airflow
datetime
pandas
sqlalchemy
requests